Fix cargo fmt
